Unify types
