+++ app/frontend/Pages/Manage/CreateExperience/CreatePoll/CreatePoll.tsx
@@ -1,8 +1,7 @@
 import { Button, TextInput } from '@cctv/core';
 import { Dropdown } from '@cctv/core/Dropdown/Dropdown';
 import { BlockStatus, ParticipantSummary } from '@cctv/types';
-
-import { BlockComponentProps, PollData } from '../types';
+import { BlockComponentProps, PollData } from '@cctv/types';
 
 import styles from './CreatePoll.module.scss';
